fix: Don't allow cyclic dependencies
This change passes `MacroExpander` to all macro expansions which allows
them to keep some state of their own between multiple macro expansions.
This was supposed to work before as well but due to some changes the
cyclic dependencies would only be detected if a file tried to include
itself and not if the import was through another file.